requirements.txt
fastapi
uvicorn
PyJWT
python-dotenv
stripe
pydantic
